requirements.txt
# Django Framework
Django==5.2
djangorestframework==3.15.2

# Imagens
Pillow==11.0.0

# Geração de PDF
reportlab==4.2.5
weasyprint==62.3

# Timezone
pytz==2024.2

# Database
# SQLite já vem com Python, mas para produção considere:
# psycopg2-binary==2.9.10  # PostgreSQL
# mysqlclient==2.2.6  # MySQL

# Utilitários
python-decouple==3.8
